Added simpler test of NCMCAlchemicalIntegrator that fails on OpenCL
