+++ perses/tests/utils.py
@@ -1,0 +1,167 @@
+"""
+Unit tests for NCMC switching engine.
+
+"""
+
+__author__ = 'John D. Chodera'
+
+################################################################################
+# IMPORTS
+################################################################################
+
+from simtk import openmm, unit
+from simtk.openmm import app
+import os, os.path
+import sys, math
+import numpy as np
+from functools import partial
+from pkg_resources import resource_filename
+from openeye import oechem
+if sys.version_info >= (3, 0):
+    from io import StringIO
+    from subprocess import getstatusoutput
+else:
+    from cStringIO import StringIO
+    from commands import getstatusoutput
+
+################################################################################
+# CONSTANTS
+################################################################################
+
+kB = unit.BOLTZMANN_CONSTANT_kB * unit.AVOGADRO_CONSTANT_NA
+temperature = 300.0 * unit.kelvin
+kT = kB * temperature
+beta = 1.0/kT
+
+################################################################################
+# UTILITIES
+################################################################################
+
+# TODO: Move some of these utility routines to openmoltools.
+
+def extractPositionsFromOEMOL(molecule):
+    positions = unit.Quantity(np.zeros([molecule.NumAtoms(), 3], np.float32), unit.angstroms)
+    coords = molecule.GetCoords()
+    for index in range(molecule.NumAtoms()):
+        positions[index,:] = unit.Quantity(coords[index], unit.angstroms)
+    return positions
+
+def createOEMolFromIUPAC(iupac_name='bosutinib'):
+    from openeye import oechem, oeiupac, oeomega
+
+    # Create molecule.
+    mol = oechem.OEMol()
+    oeiupac.OEParseIUPACName(mol, iupac_name)
+    mol.SetTitle(iupac_name)
+
+    # Assign aromaticity and hydrogens.
+    oechem.OEAssignAromaticFlags(mol, oechem.OEAroModelOpenEye)
+    oechem.OEAddExplicitHydrogens(mol)
+
+    # Create atom names.
+    oechem.OETriposAtomNames(mol)
+
+    # Assign geometry
+    omega = oeomega.OEOmega()
+    omega.SetMaxConfs(1)
+    omega.SetIncludeInput(False)
+    omega.SetStrictStereo(True)
+    omega(mol)
+
+    return mol
+
+def generate_gaff_xml():
+    """
+    Return a file-like object for `gaff.xml`
+    """
+    from openmoltools import amber
+    gaff_dat_filename = amber.find_gaff_dat()
+
+    # Generate ffxml file contents for parmchk-generated frcmod output.
+    leaprc = StringIO("parm = loadamberparams %s" % gaff_dat_filename)
+    import parmed
+    params = parmed.amber.AmberParameterSet.from_leaprc(leaprc)
+    params = parmed.openmm.OpenMMParameterSet.from_parameterset(params)
+    citations = """\
+Wang, J., Wang, W., Kollman P. A.; Case, D. A. "Automatic atom type and bond type perception in molecular mechanical calculations". Journal of Molecular Graphics and Modelling , 25, 2006, 247260.
+Wang, J., Wolf, R. M.; Caldwell, J. W.;Kollman, P. A.; Case, D. A. "Development and testing of a general AMBER force field". Journal of Computational Chemistry, 25, 2004, 1157-1174.
+"""
+    ffxml = str()
+    gaff_xml = StringIO(ffxml)
+    provenance=dict(OriginalFile='gaff.dat', Reference=citations)
+    params.write(gaff_xml, provenance=provenance)
+
+    return gaff_xml
+
+def get_data_filename(relative_path):
+    """Get the full path to one of the reference files shipped for testing
+
+    In the source distribution, these files are in ``perses/data/*/``,
+    but on installation, they're moved to somewhere in the user's python
+    site-packages directory.
+
+    Parameters
+    ----------
+    name : str
+        Name of the file to load (with respect to the openmoltools folder).
+
+    """
+
+    fn = resource_filename('perses', relative_path)
+
+    if not os.path.exists(fn):
+        raise ValueError("Sorry! %s does not exist. If you just added it, you'll have to re-install" % fn)
+
+    return fn
+
+def createSystemFromIUPAC(iupac_name):
+    """
+    Create an openmm system out of an oemol
+
+    Parameters
+    ----------
+    iupac_name : str
+        IUPAC name
+
+    Returns
+    -------
+    molecule : openeye.OEMol
+        OEMol molecule
+    system : openmm.System object
+        OpenMM system
+    positions : [n,3] np.array of floats
+        Positions
+    topology : openmm.app.Topology object
+        Topology
+    """
+
+    # Create OEMol
+    molecule = createOEMolFromIUPAC(iupac_name)
+
+    # Generate a topology.
+    from openmoltools.forcefield_generators import generateTopologyFromOEMol
+    topology = generateTopologyFromOEMol(molecule)
+
+    # Initialize a forcefield with GAFF.
+    # TODO: Fix path for `gaff.xml` since it is not yet distributed with OpenMM
+    from simtk.openmm.app import ForceField
+    gaff_xml_filename = get_data_filename('data/gaff.xml')
+    forcefield = ForceField(gaff_xml_filename)
+
+    # Generate template and parameters.
+    from openmoltools.forcefield_generators import generateResidueTemplate
+    [template, ffxml] = generateResidueTemplate(molecule)
+
+    # Register the template.
+    forcefield.registerResidueTemplate(template)
+
+    # Add the parameters.
+    forcefield.loadFile(StringIO(ffxml))
+
+    # Create the system.
+    system = forcefield.createSystem(topology)
+
+    # Extract positions
+    positions = extractPositionsFromOEMOL(molecule)
+
+    return (molecule, system, positions, topology)
